Move ThreadPoolUtils to new package
