+++ agent/agent-plugins/jdk-thread/src/main/java/org/bithon/agent/plugin/thread/interceptor/ForkJoinPool$Ctor.java
@@ -18,9 +18,9 @@ package org.bithon.agent.plugin.thread.interceptor;
 
 import org.bithon.agent.bootstrap.aop.AbstractInterceptor;
 import org.bithon.agent.bootstrap.aop.AopContext;
-import org.bithon.agent.plugin.thread.ThreadPoolUtils;
 import org.bithon.agent.plugin.thread.metrics.ForkJoinPoolMetrics;
 import org.bithon.agent.plugin.thread.metrics.ThreadPoolMetricRegistry;
+import org.bithon.agent.plugin.thread.utils.ThreadPoolUtils;
 import org.bithon.component.commons.utils.ReflectionUtils;
 
 import java.util.concurrent.ForkJoinPool;
